move pattern in instanceof (UniPatternInstanceOf) in java 15
