+++ api/src/main/java/org/galaxy/uniflow/api/expressions/UniInstanceOf.java
@@ -1,9 +1,7 @@
 package org.galaxy.uniflow.api.expressions;
 
 import org.galaxy.uniflow.api.UniElement;
-import org.galaxy.uniflow.api.pattern.UniPattern;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 public interface UniInstanceOf extends UniExpression {
 
@@ -12,9 +10,5 @@ public interface UniInstanceOf extends UniExpression {
     @NotNull UniExpression getExpression();
 
     @NotNull UniElement getType();
-
-    void setPattern(@Nullable UniPattern pattern);
-
-    @Nullable UniPattern getPattern();
 
 }
