+++ javac/src/main/java/org/galaxy/uniflow/javac/expression/JavacInstanceOf.java
@@ -4,11 +4,9 @@ import com.sun.tools.javac.tree.JCTree;
 import org.galaxy.uniflow.api.UniElement;
 import org.galaxy.uniflow.api.expressions.UniExpression;
 import org.galaxy.uniflow.api.expressions.UniInstanceOf;
-import org.galaxy.uniflow.api.pattern.UniPattern;
 import org.galaxy.uniflow.javac.util.JavacUnwrapper;
 import org.galaxy.uniflow.javac.util.UniflowWrapper;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 public class JavacInstanceOf extends JavacExpression<JCTree.JCInstanceOf> implements UniInstanceOf {
 
@@ -29,15 +27,5 @@ public class JavacInstanceOf extends JavacExpression<JCTree.JCInstanceOf> implem
     @Override
     public @NotNull UniElement getType() {
         return UniflowWrapper.wrap(tree.getType());
-    }
-
-    @Override
-    public void setPattern(@Nullable UniPattern pattern) {
-        tree.pattern = JavacUnwrapper.unwrap(pattern);
-    }
-
-    @Override
-    public @Nullable UniPattern getPattern() {
-        return (UniPattern) UniflowWrapper.wrap(tree.pattern);
     }
 }
